Split up test cases into individual files
