+++ tests/FunctionAwaitRaceTest_.php
@@ -1,0 +1,1 @@
+<?php
