+++ tests/FunctionAwaitTest.php
@@ -1,0 +1,27 @@
+<?php
+
+class FunctionAwaitTest extends TestCase
+{
+    public function testAwaitOneRejected()
+    {
+        $promise = $this->createPromiseRejected(new Exception('test'));
+
+        $this->setExpectedException('Exception', 'test');
+        $this->block->await($promise);
+    }
+
+    public function testAwaitOneResolved()
+    {
+        $promise = $this->createPromiseResolved(2);
+
+        $this->assertEquals(2, $this->block->await($promise));
+    }
+
+    public function testAwaitOneInterrupted()
+    {
+        $promise = $this->createPromiseResolved(2, 0.02);
+        $this->createTimerInterrupt(0.01);
+
+        $this->assertEquals(2, $this->block->await($promise));
+    }
+}
